+++ nata-monkey/src/main/java/com/nata/runners/AllRunner.java
@@ -1,7 +1,6 @@
-package com.nata;
+package com.nata.runners;
 
 import com.nata.cmd.AdbDevice;
-import com.nata.monkeys.AbstractMonkey;
 import com.nata.monkeys.QLearningMonkey;
 import com.nata.monkeys.RandomMonkey;
 
@@ -10,18 +9,18 @@ import com.nata.monkeys.RandomMonkey;
  * Blog: mclspace.com  Email: rdmclin2@gamil.com
  * Update: 2016-01-19 16:15
  */
-public class QLMRunner {
+public class AllRunner {
     public static void main(String[] args) {
-//      String pkg = "com.zhihu.android";
-//      String act = ".app.ui.activity.MainActivity";
         String pkg = "com.cvicse.zhnt";
         String act = ".LoadingActivity";
 
+        //Set Monkey settings
         AdbDevice device = new AdbDevice();
-        AbstractMonkey qLearningMonkey = new QLearningMonkey(pkg, act, device);
-//      qLearningMonkey.learn("rules/zhihu.json");
+        RandomMonkey randomMonkey = new RandomMonkey(pkg, act, device);
+        QLearningMonkey qLearningMonkey = new QLearningMonkey(pkg, act, device);
+        randomMonkey.play();
+        randomMonkey.report();
         qLearningMonkey.play();
         qLearningMonkey.report();
-
     }
 }