+++ nata-monkey/src/main/java/com/nata/runners/QLMRunner.java
@@ -1,6 +1,8 @@
-package com.nata;
+package com.nata.runners;
 
 import com.nata.cmd.AdbDevice;
+import com.nata.monkeys.AbstractMonkey;
+import com.nata.monkeys.QLearningMonkey;
 import com.nata.monkeys.RandomMonkey;
 
 /**
@@ -8,16 +10,18 @@ import com.nata.monkeys.RandomMonkey;
  * Blog: mclspace.com  Email: rdmclin2@gamil.com
  * Update: 2016-01-19 16:15
  */
-public class RMRunner {
+public class QLMRunner {
     public static void main(String[] args) {
 //      String pkg = "com.zhihu.android";
 //      String act = ".app.ui.activity.MainActivity";
-
         String pkg = "com.cvicse.zhnt";
         String act = ".LoadingActivity";
+
         AdbDevice device = new AdbDevice();
-        RandomMonkey randomMonkey = new RandomMonkey(pkg, act, device);
-        randomMonkey.play();
-        randomMonkey.report();
+        AbstractMonkey qLearningMonkey = new QLearningMonkey(pkg, act, device);
+//      qLearningMonkey.learn("rules/zhihu.json");
+        qLearningMonkey.play();
+        qLearningMonkey.report();
+
     }
 }