+++ nata-monkey/src/main/java/com/nata/runners/RMRunner.java
@@ -1,7 +1,6 @@
-package com.nata;
+package com.nata.runners;
 
 import com.nata.cmd.AdbDevice;
-import com.nata.monkeys.QLearningMonkey;
 import com.nata.monkeys.RandomMonkey;
 
 /**
@@ -9,18 +8,16 @@ import com.nata.monkeys.RandomMonkey;
  * Blog: mclspace.com  Email: rdmclin2@gamil.com
  * Update: 2016-01-19 16:15
  */
-public class AllRunner {
+public class RMRunner {
     public static void main(String[] args) {
+//      String pkg = "com.zhihu.android";
+//      String act = ".app.ui.activity.MainActivity";
+
         String pkg = "com.cvicse.zhnt";
         String act = ".LoadingActivity";
-
-        //Set Monkey settings
         AdbDevice device = new AdbDevice();
         RandomMonkey randomMonkey = new RandomMonkey(pkg, act, device);
-        QLearningMonkey qLearningMonkey = new QLearningMonkey(pkg, act, device);
         randomMonkey.play();
         randomMonkey.report();
-        qLearningMonkey.play();
-        qLearningMonkey.report();
     }
 }